requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.9.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
dj-database-url==2.2.0
Django==5.0.7
django-anymail==13.1
django-axes==8.0.0
django-csp==4.0
django-debug-toolbar==6.0.0
django-ratelimit==4.1.0
django-sslserver==0.22
django-widget-tweaks==1.5.0
gunicorn==22.0.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.9
pycparser==2.23
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.0.1
requests==2.32.5
sentry-sdk==2.37.1
sqlparse==0.5.3
typing_extensions==4.15.0
urllib3==2.5.0
whitenoise==6.7.0
